+++ core/ui/postcard/build.gradle.kts
@@ -1,33 +1,33 @@
 plugins {
     id("org.mozilla.social.android.library")
     id("org.mozilla.social.android.library.compose")
-    alias(libs.plugins.kotlin.ksp)
 }
 
 android {
-    namespace = "org.mozilla.social.feature.thread"
+    namespace = "org.mozilla.social.core.ui.postcard"
 }
 
 dependencies {
     implementation(project(":core:model"))
-    implementation(project(":core:domain"))
-    implementation(project(":core:data"))
-    implementation(project(":core:datastore"))
     implementation(project(":core:designsystem"))
-    implementation(project(":core:ui:common"))
-    implementation(project(":core:ui:postcard"))
     implementation(project(":core:common"))
+    implementation(project(":core:data"))
     implementation(project(":core:navigation"))
+    implementation(project(":core:ui:common"))
 
-    implementation(libs.androidx.paging.runtime)
+    implementation(libs.androidx.paging.compose)
 
     implementation(libs.androidx.navigation.compose)
     implementation(libs.koin.core)
     implementation(libs.koin.androidx.compose)
 
-    implementation(libs.androidx.datastore)
-    implementation(libs.protobuf.kotlin.lite)
+    implementation(libs.kotlinx.datetime)
+
+    implementation(libs.coil)
+    implementation(libs.androidx.media3.exoplayer)
+    implementation(libs.androidx.media3.ui)
 
     implementation(libs.jakewharton.timber)
-    implementation(libs.androidx.navigation.compose)
+
+    implementation(libs.androidx.browser)
 }
