+++ core/ui/postcard/src/main/java/org/mozilla/social/core/ui/postcard/PostCardList.kt
@@ -1,4 +1,4 @@
-package org.mozilla.social.core.ui.common.postcard
+package org.mozilla.social.core.ui.postcard
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
