+++ core/ui/postcard/src/main/java/org/mozilla/social/core/ui/postcard/poll/PollInteractions.kt
@@ -1,4 +1,4 @@
-package org.mozilla.social.core.ui.common.poll
+package org.mozilla.social.core.ui.postcard.poll
 
 interface PollInteractions {
     fun onVoteClicked(pollId: String, choices: List<Int>) = Unit
