+++ core/ui/postcard/src/main/java/org/mozilla/social/core/ui/postcard/poll/PollMapping.kt
@@ -1,8 +1,8 @@
-package org.mozilla.social.core.ui.common.poll
+package org.mozilla.social.core.ui.postcard.poll
 
 import org.mozilla.social.common.utils.StringFactory
 import org.mozilla.social.common.utils.timeLeft
-import org.mozilla.social.core.ui.common.R
+import org.mozilla.social.core.ui.postcard.R
 import org.mozilla.social.model.Poll
 import org.mozilla.social.model.PollOption
