+++ core/ui/postcard/src/main/res/values/strings.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="vote_button">Vote</string>
+
+    <plurals name="vote_count_and_percent">
+        <item quantity="one">%1$d vote - %2$d%%</item>
+        <item quantity="other">%1$d votes - %2$d%%</item>
+    </plurals>
+
+    <plurals name="vote_count">
+        <item quantity="one">%d vote</item>
+        <item quantity="other">%d votes</item>
+    </plurals>
+
+    <string name="poll_closed">poll closed</string>
+    <string name="poll_choose_one_or_more_options">Choose one or more</string>
+</resources>
