+++ feature/account/build.gradle.kts
@@ -15,6 +15,7 @@ dependencies {
     implementation(project(":core:data"))
     implementation(project(":core:designsystem"))
     implementation(project(":core:ui:common"))
+    implementation(project(":core:ui:postcard"))
     implementation(project(":core:common"))
     implementation(project(":core:database"))
     implementation(project(":core:navigation"))
