+++ settings.gradle.kts
@@ -50,3 +50,5 @@ include(":core:analytics")
 include(":feature:followers")
 include(":core:navigation")
 include(":feature:discover")
+include(":core:ui:postcard")
+include(":core:ui:poll")
